+++ src/main/java/com/jug/util/ArgbDrawingUtils.java
@@ -23,24 +23,28 @@ public class ArgbDrawingUtils {
      *
      * @param imgDestination image to draw calculated overlay pixel values to
      * @param imgSource      pixel value source
-     * @param segments       a <code>List</code> of the hypotheses containing
+     * @param hypotheses       a <code>List</code> of the hypotheses containing
      *                       component-tree-nodes that represent the optimal segmentation
      *                       (the one returned by the solution to the ILP)
      */
-    public static void drawSegments(final Img<ARGBType> imgDestination, final Img<ARGBType> imgSource, final long offsetX, final long offsetY, final Iterable<Hypothesis<AdvancedComponent<FloatType>>> segments) {
+    public static void drawSegments(final Img<ARGBType> imgDestination, final Img<ARGBType> imgSource, final long offsetX, final long offsetY, final Iterable<Hypothesis<AdvancedComponent<FloatType>>> hypotheses) {
         final RandomAccess<ARGBType> targetImage = imgDestination.randomAccess();
         final RandomAccess<ARGBType> sourceImage = imgSource.randomAccess();
-        for (final Hypothesis<AdvancedComponent<FloatType>> hypothesis : segments) {
+        for (final Hypothesis<AdvancedComponent<FloatType>> hypothesis : hypotheses) {
             final AdvancedComponent<FloatType> component = hypothesis.getWrappedComponent();
             Function<Integer, ARGBType> pixelOverlayColorCalculator;
-            if (hypothesis.isPruned()) {
+            if (hypothesis.isSelected()) {
+                pixelOverlayColorCalculator = grayscaleValue -> calculateBluePixelOverlayValue(grayscaleValue); /* highlight selected component in blue */
+            } else if (hypothesis.isPruned()) {
                 pixelOverlayColorCalculator = grayscaleValue -> calculateGrayPixelOverlayValue(grayscaleValue); /* highlight pruned component in gray */
             } else if (hypothesis.isForced()) {
                 pixelOverlayColorCalculator = grayscaleValue -> calculateYellowPixelOverlayValue(grayscaleValue); /* highlight enforced component in yellow */
             } else if (hypothesis.isForceIgnored()) {
                 pixelOverlayColorCalculator = grayscaleValue -> calculateRedPixelOverlayValue(grayscaleValue); /* highlight enforced component in yellow */
-            } else {
+            } else if (hypothesis.isActive()) {
                 pixelOverlayColorCalculator = grayscaleValue -> calculateGreenPixelOverlayValue(grayscaleValue); /* highlight optimal component in green */
+            } else {
+                continue; /* do not draw segments that are inactive and not force-ignored */
             }
             drawSegmentColorOverlay(component, targetImage, sourceImage, offsetX, offsetY, pixelOverlayColorCalculator);
             if (!hypothesis.labels.isEmpty()) {